[#68] fix: readme 수정 및 Controller 수정
